feat: show last 24hrs trips
